[feat] 소통방 추가
