+++ src/constants/animation.ts
@@ -1,0 +1,8 @@
+const ANIMATION = {
+  FADE_IN_LEFT: 'animate-fade-in-to-left',
+  FADE_OUT_LEFT: 'animate-fade-out-to-left',
+  FADE_IN_RIGHT: 'animate-fade-in-to-right',
+  FADE_OUT_RIGHT: 'animate-fade-out-to-right',
+};
+
+export { ANIMATION };
