+++ src/utils/formatter.ts
@@ -1,0 +1,5 @@
+const numberFormatter = new Intl.NumberFormat('ko-kr', {
+  notation: 'standard',
+}).format;
+
+export { numberFormatter };
